Consolidate JournalEntryFormPage to just be the new entry form
